Multi-PDFs_ChatApp_AI-Agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langchain-classic
PyPDF2
faiss-cpu
python-multipart